+++ app/com/mohiva/play/silhouette/impl/providers/OAuth2Provider.scala
@@ -21,18 +21,17 @@ package com.mohiva.play.silhouette.impl.providers
 
 import java.net.URLEncoder._
 
-import com.mohiva.play.silhouette._
 import com.mohiva.play.silhouette.api._
 import com.mohiva.play.silhouette.api.exceptions._
 import com.mohiva.play.silhouette.api.services.AuthInfo
-import com.mohiva.play.silhouette.api.util.HTTPLayer
-import com.mohiva.play.silhouette.impl
+import com.mohiva.play.silhouette.api.util.{ ExtractableRequest, HTTPLayer }
+import com.mohiva.play.silhouette.{ impl, _ }
 import com.mohiva.play.silhouette.impl.providers.OAuth2Provider._
 import play.api.libs.concurrent.Execution.Implicits._
 import play.api.libs.functional.syntax._
 import play.api.libs.json._
 import play.api.libs.ws.WSResponse
-import play.api.mvc.{ RequestHeader, Result, Results }
+import play.api.mvc._
 
 import scala.concurrent.Future
 import scala.util.{ Failure, Success, Try }
@@ -91,17 +90,16 @@ abstract class OAuth2Provider(httpLayer: HTTPLayer, stateProvider: OAuth2StatePr
   /**
    * Starts the authentication process.
    *
-   * @param request The request header.
+   * @param request The request.
    * @return Either a Result or the auth info from the provider.
    */
-  def authenticate()(implicit request: RequestHeader): Future[Either[Result, OAuth2Info]] = {
-    logger.debug("[Silhouette][%s] Query string: %s".format(id, request.rawQueryString))
-    request.queryString.get(Error).flatMap(_.headOption).map {
+  def authenticate[B]()(implicit request: ExtractableRequest[B]): Future[Either[Result, OAuth2Info]] = {
+    request.extractString(Error).map {
       case e @ AccessDenied => new AccessDeniedException(AuthorizationError.format(id, e))
       case e => new AuthenticationException(AuthorizationError.format(id, e))
     } match {
       case Some(throwable) => Future.failed(throwable)
-      case None => request.queryString.get(Code).flatMap(_.headOption) match {
+      case None => request.extractString(Code) match {
         // We're being redirected back from the authorization server with the access code
         case Some(code) => stateProvider.validate(id).recoverWith {
           case e => Future.failed(new AuthenticationException(InvalidState.format(id), e))
@@ -238,20 +236,22 @@ trait OAuth2StateProvider {
    * Validates the provider and the client state.
    *
    * @param id The provider ID.
-   * @param request The request header.
+   * @param request The request.
+   * @tparam B The type of the request body.
    * @return The state on success, otherwise an failure.
    */
-  def validate(id: String)(implicit request: RequestHeader): Future[State]
+  def validate[B](id: String)(implicit request: ExtractableRequest[B]): Future[State]
 
   /**
    * Publishes the state to the client.
    *
    * @param result The result to send to the client.
    * @param state The state to publish.
-   * @param request The request header.
+   * @param request The request.
+   * @tparam B The type of the request body.
    * @return The result to send to the client.
    */
-  def publish(result: Result, state: State)(implicit request: RequestHeader): Result
+  def publish[B](result: Result, state: State)(implicit request: ExtractableRequest[B]): Result
 }
 
 /**